cleanup proxy/envoy package
