+++ pilot/cmd/pilot-agent/main_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/onsi/gomega"
 
 	"istio.io/istio/pilot/pkg/model"
-	"istio.io/istio/pilot/pkg/proxy/envoy"
+	"istio.io/istio/pilot/pkg/security/authn/utils"
 	"istio.io/istio/pilot/pkg/serviceregistry"
 	"istio.io/istio/pkg/config/constants"
 )
@@ -77,7 +77,7 @@ func TestCustomMixerSanIfAuthenticationMutualDomainKubernetes(t *testing.T) {
 	registryID = serviceregistry.Kubernetes
 
 	setSpiffeTrustDomain("", role.DNSDomain)
-	mixerSAN := envoy.GetSAN("", mixerIdentity)
+	mixerSAN := utils.GetSAN("", mixerIdentity)
 
 	g.Expect(mixerSAN).To(gomega.Equal("spiffe://mesh.com/mixer-identity"))
 }
