--- pilot/pkg/proxy/envoy/testdata/bootstrap.json
@@ -1,1 +1,0 @@
-{"key": "value"}
