+++ pilot/pkg/security/authn/v1beta1/policy_applier_test.go
@@ -39,7 +39,7 @@ import (
 	"istio.io/istio/pilot/pkg/model/test"
 	"istio.io/istio/pilot/pkg/networking"
 	pilotutil "istio.io/istio/pilot/pkg/networking/util"
-	xdsfilters "istio.io/istio/pilot/pkg/proxy/envoy/filters"
+	"istio.io/istio/pilot/pkg/xds"
 	protovalue "istio.io/istio/pkg/proto"
 	authn_alpha "istio.io/istio/security/proto/authentication/v1alpha1"
 	authn_filter "istio.io/istio/security/proto/envoy/config/filter/http/authn/v2alpha1"
@@ -1255,7 +1255,7 @@ func TestOnInboundFilterChain(t *testing.T) {
 				ApplicationProtocols: []string{"istio-peer-exchange", "istio"},
 			},
 			ListenerFilters: []*listener.ListenerFilter{
-				xdsfilters.TLSInspector,
+				xds.TLSInspector,
 			},
 		},
 		{
