+++ pilot/pkg/xds/simple.go
